PoolNet update

+++ PoolNet/process_gazefollow.py
@@ -20,6 +20,7 @@ channel = 1
 show = False
 save = True
 dockerPath = True
+minHW = 112
 
 # path
 home = os.path.expanduser("~")
@@ -36,7 +37,8 @@ elif mode == "train":
         load_dir = "/exper/gaze/attention-target-detection/data/gazefollow/train"
         save_dir = "/exper/gaze/attention-target-detection/data/gazefollow/train_sal"
 
-os.mkdir(save_dir)
+if not os.path.exists(save_dir):
+    os.mkdir(save_dir)
 
 pretraind_model = "/exper/salient/PoolNet/dataset/run-0/run-0/models/final.pth" if dockerPath else os.path.join(home, "exper/salient/PoolNet/dataset/run-0/run-0/models/final.pth")
 print('Loading pre-trained model from %s...' % pretraind_model)
@@ -49,31 +51,47 @@ net.to(device)
 net.eval()
 
 for foldername in os.listdir(load_dir):
-    print("foldername: ", foldername)
     load_foldername = os.path.join(load_dir, foldername)
     save_foldername = os.path.join(save_dir, foldername)
-    os.mkdir(save_foldername)
-    for filename in os.listdir(load_foldername):
-        load_filename = os.path.join(load_foldername, filename)
-        save_filename = os.path.join(save_foldername, filename)
+    if not os.path.exists(save_foldername):
+        print("foldername: ", foldername)
+        os.mkdir(save_foldername)
+        for filename in os.listdir(load_foldername):
+            # print("    filename: ", filename)
+            load_filename = os.path.join(load_foldername, filename)
+            save_filename = os.path.join(save_foldername, filename)
 
-        # load img
-        img = cv2.imread(load_filename)
-        img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
-        img = img.transpose((2, 0, 1))
-        img = torch.from_numpy(img)
-        images = Variable(img).unsqueeze(0).type(torch.FloatTensor)
-        if torch.cuda.is_available():
-            images = images.to(device)
+            # load img
+            img = cv2.imread(load_filename)
+            img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
+            height, width, channels = img.shape
+            m = min(height, width)
+            if m < minHW:
+                if height < width:
+                    scale = minHW / height
+                    img = cv2.resize(img, (int(width * scale), minHW), interpolation=cv2.INTER_CUBIC)
+                else:
+                    scale = minHW / width
+                    img = cv2.resize(img, (minHW, int(height * scale)), interpolation=cv2.INTER_CUBIC)
 
-        # predict
-        with torch.no_grad():
-            preds = net(images)
-            pred = np.squeeze(torch.sigmoid(preds).cpu().data.numpy())
-            multi_fuse = 255 * pred
+            img = img.transpose((2, 0, 1))
+            img = torch.from_numpy(img)
+            images = Variable(img).unsqueeze(0).type(torch.FloatTensor)
+            if torch.cuda.is_available():
+                images = images.to(device)
 
-        if save:
-            cv2.imwrite(save_filename, multi_fuse)
-        if show:
-            cv2.imshow(filename, multi_fuse)
-            cv2.waitKey(0)
+            # predict
+            with torch.no_grad():
+                preds = net(images)
+                pred = np.squeeze(torch.sigmoid(preds).cpu().data.numpy())
+                multi_fuse = 255 * pred
+
+            if save:
+                if m < minHW:
+                    multi_fuse = cv2.resize(multi_fuse, (width, height), interpolation=cv2.INTER_CUBIC)
+                cv2.imwrite(save_filename, multi_fuse)
+            if show:
+                cv2.imshow(filename, multi_fuse)
+                cv2.waitKey(0)
+    else:
+        print("foldername %s exist.", foldername)
